@@ -3,14 +3,16 @@ description: A new flutter module project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  flutter_screenutil: ^4.0.3+1
+  flutter_screenutil: ^5.0.0+2
+  flutter_inappwebview: ^5.3.2
+  kraken: ">= 0.0.1"
 
 dev_dependencies:
   flutter_test:
@@ -46,4 +48,4 @@ flutter:
   module:
     androidX: true
     androidPackage: com.example.flutter_module
-    iosBundleIdentifier: com.example.flutterModule
+    iosBundleIdentifier: com.itfenbao.ios.flutterModule
