@@ -5,13 +5,13 @@ UniappMethodChannel uniapp = UniappMethodChannel();
 
 class GLObserver extends NavigatorObserver {
   @override
-  void didPush(Route route, Route previousRoute) {
+  void didPush(Route route, Route? previousRoute) {
     super.didPush(route, previousRoute);
     uniapp.fireEvent("canPop", true);
   }
 
   @override
-  void didPop(Route route, Route previousRoute) {
+  void didPop(Route route, Route? previousRoute) {
     super.didPop(route, previousRoute);
     Future.delayed(const Duration(milliseconds: 10), () {
       var canPop = Navigator.canPop(uniapp.getContext());
